pyproject.toml
[tool.poetry]
name = "poetry_template"
version = "0.1.0"
description = ""
authors = ["Research Computing Service, Imperial College London <rcs-support@imperial.ac.uk>"]
readme = "README.md"
homepage = "https://github.com/ImperialCollegeLondon/poetry_template"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

msgspec = "^0.13.1"
strawberry-graphql = "^0.167.1"

[tool.poetry.dev-dependencies]



[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

